refacto appView+ remove createbord folder provider dep
