+++ frontend/src/components/app-view/index.scss
