+++ frontend/src/routes/board/index.tsx
@@ -1,15 +1,12 @@
 import { BoardView } from "~/components/board-view/BoardView";
 import { BoardProvider } from "~/providers/BoardProvider";
-import { MediaLibraryProvider } from "~/providers/MediaLibraryProvider";
 import { SVGProvider } from "~/providers/SVGProvider";
 
 export const App = () => {
   return (
     <BoardProvider>
       <SVGProvider>
-        <MediaLibraryProvider>
           <BoardView />
-        </MediaLibraryProvider>
       </SVGProvider>
     </BoardProvider>
   );
